+++ samoJ/src/samoJ/Circle.java
@@ -2,7 +2,7 @@ package samoJ;
 
 public class Circle extends Shape {
 	Coord theCenter;
-	int Radius;
+	double Radius;
 
 	/**
 	 * 
@@ -13,14 +13,14 @@ public class Circle extends Shape {
 	 * @param z1
 	 */
 
-	public Circle(int r, int x1, int y1, int z1) {
+	public Circle(double r, double x1, double y1, double z1) {
 		theCenter = new Coord(x1, y1, z1);
 		Radius = r;
 		create_circle();
 	}
 
-	public Circle(int rx, int ry, int rz, int x2, int y2, int z2) {
-		this((int)Math.round(Math.sqrt((rx - x2) * (rx - x2) +(ry - y2)*(ry - y2))), x2, y2, z2);
+	public Circle(double rx, double ry, double rz, double x2, double y2, double z2) {
+		this((double)Math.round(Math.sqrt((rx - x2) * (rx - x2) +(ry - y2)*(ry - y2))), x2, y2, z2);
 		/*int xd = rx - x2;
 		int yd = ry - y2;
 		int r = (int) Math.sqrt(xd * xd + yd * yd);
@@ -30,9 +30,9 @@ public class Circle extends Shape {
 	}
 
 	public void create_circle() {
-		// int[] lines = circle_lines(Radius, theCenter.x, theCenter.y); красота!!!
+		// int[] lines = circle_lines(Radius, theCenter.x, theCenter.y); пїЅпїЅпїЅпїЅпїЅпїЅпїЅ!!!
 
-		int[] points = circle_points();
+		double[] points = circle_points();
 
 		for (int i = 0; i < points.length - 2; i += 2) {
 
@@ -44,7 +44,7 @@ public class Circle extends Shape {
 				points[points.length - 1], 0, points[0], points[1], 0));
 
 		// Crossers in center of circle
-		int s = Radius / 20;
+		double s = Radius / 20;
 		add(new PrimitiveLine(theCenter.x - s, theCenter.y, 0, theCenter.x + s, theCenter.y, 0));
 		add(new PrimitiveLine(theCenter.x, theCenter.y - s, 0, theCenter.x, theCenter.y + s, 0));
 
@@ -58,17 +58,17 @@ public class Circle extends Shape {
 	 * @param y
 	 * @return
 	 */
-	public int[] circle_points() {
+	public double[] circle_points() {
 		int segments = core.Values.circle_segments;
-		int[] lines = new int[segments * 2];
+		double[] lines = new double[segments * 2];
 		// int w, h = 2.0f*Radius;
 		double angle_increment = (double) Math.PI * 2.0f / segments;
 		double theta = 0;
 
 		for (int i = 0; i < segments; i++) {
 
-			lines[i * 2] = (int) (Radius * Math.cos(theta) + theCenter.x);
-			lines[i * 2 + 1] = (int) (Radius * Math.sin(theta) + theCenter.y);
+			lines[i * 2] = (double) (Radius * Math.cos(theta) + theCenter.x);
+			lines[i * 2 + 1] = (double) (Radius * Math.sin(theta) + theCenter.y);
 
 			theta += angle_increment;
 		}
